backend/requirements.txt
fastapi
uvicorn
tensorflow
torch
torchvision
pillow
numpy
opencv-python
mediapipe
python-multipart